Create 0_basic_clam__with_pain.py
Experimentation and testing of pain control neurons in Arch class and compatibility with core, for the Pavlov app by @Rafipilot

Note: I've not yet tested that this works technically (the way to test it would be to complete a simple pain function and then test it in the clam, to see that the signals are passing through properly from Arch to Agent). Testing that this works functionally is application dependent, the next step (what you'll do in the Pavlov app).