optimized the 6th parameter.
@spolisar 
Now, the 6th parameter in *rectangular* and *nearest neighbor* must always be an integer. Previously, we had to explicitly define `False` if we didn’t want a random connection, but now the default is `False`. Therefore, the 6th parameter must always be an integer, and your error message will make more sense now.