Major Qaux bug fixes
Before merging qaux, I checked both core and arch and there are serious errors that meant the feature wasn't working at all when we tested it.

Let's re-test it technically (run the code, try it) and functionally (by updating it in the Pavlov app; what we tried before was not an accurate test).

Co-Authored-By: Rafayel Latif <74052861+Rafipilot@users.noreply.github.com>